+++ Basil/include/Renderer/Renderer2D.h
@@ -1,0 +1,20 @@
+/*
+ * This header file declares the 2D renderer class.
+ */
+#pragma once
+
+#include "Renderer/OrthographicCamera.h"
+
+namespace Basil
+{
+	class Renderer2D
+	{
+		public:
+			static void init();
+			static void shutdown();
+			static void beginScene(const OrthographicCamera& camera);
+			static void endScene();
+			static void drawQuad(const glm::vec2& position, const glm::vec2& size, const glm::vec4& color);
+			static void drawQuad(const glm::vec3& position, const glm::vec2& size, const glm::vec4& color);
+	};
+}